requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
